Добавляет debounce для меток карты
