+++ js/main.js
@@ -1,4 +1,5 @@
 import './form-validation.js';
+import { debounce } from './util.js';
 import { renderMap, renderPinGroup } from './map.js';
 import { getData } from './api.js';
 import { setAdFormSubmit } from './form-data-send.js';
@@ -8,7 +9,7 @@ renderMap();
 
 getData((adList) => {
   renderPinGroup(adList);
-  checkFilterChange(() => renderPinGroup(adList));
+  checkFilterChange(debounce(() => renderPinGroup(adList)));
 });
 
 setAdFormSubmit();
