+++ js/util.js
@@ -27,4 +27,13 @@ const getCorrectRoomsWord = (roomsNumber) => {
 
 const isEscapeEvent = (evt) => evt.key === 'Escape' || evt.key === 'Esc';
 
-export { getCorrectGuestsWord, getCorrectRoomsWord, isEscapeEvent };
+function debounce (callback, timeoutDelay = 500) {
+  let timeoutId;
+
+  return (...rest) => {
+    clearTimeout(timeoutId);
+    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
+  };
+}
+
+export { getCorrectGuestsWord, getCorrectRoomsWord, isEscapeEvent, debounce };
